Massive Rework (Consumables)
Consumables now exist alongside items. Entities now tick, which in the future will be used to manage buffs, right now is just an empty functionality. Players always start with an item to get rid of null reference problems. Items and other things can now be put back into scenes and not just duplicated into player (this was an error resulting from the initial world item population logic).

Attack system reworked. Accuracy is now a crit determination factor instead of hits. This makes the combat feel better instead of missing all the time.

Game flow reworked to fit in with the tick function at the end. The game chooses between a combat turn or a world turn.